complete proj1b

+++ proj1b/OffByOne.java
@@ -1,5 +1,5 @@
 public class OffByOne implements CharacterComparator {
     public boolean equalChars(char x, char y) {
-        return Math.abs((int) x - y) == 1;
+        return Math.abs(x - y) == 1;
     }
 }
